Show the seach control on the embed form
