+++ app/components/companion_windows_component.html.erb
@@ -80,7 +80,7 @@
     <% end %>
 
     <%= component.with_main do %>
-      <%= render Media::EmbedThisFormComponent.new(viewer:) %>
+      <%= render EmbedThisFormComponent.new(viewer:) %>
     <% end %>
 
     <%= component.with_footer do %>
