+++ app/components/embed_this_form_component.html.erb
@@ -31,6 +31,20 @@
       <label for='sul-embed-embed-title'>title (<%= purl_object_title %>)</label>
     </div>
 
+    <% if file? %>
+      <div class='form-check'>
+        <% checkbox_options = { type: 'checkbox', id: 'sul-embed-embed-search', 'data-embed-attr': 'hide_search' } %>
+        <% checkbox_options[:checked] = 'checked' unless embed_request.hide_search? %>
+        <%= tag :input, checkbox_options %>
+        <label for='sul-embed-embed-search'>add search box</label>
+        <label for='sul-embed-min_files_to_search'>
+          for
+          <%= tag :input, type: 'text', id: 'sul-embed-min_files_to_search', 'data-embed-attr': 'min_files_to_search', value: viewer.min_files_to_search %>
+          or more files
+        </label>
+      </div>
+    <% end %>
+
     <div class='form-check'>
       <input
         type='checkbox'
